% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_factory_line.R
\name{factory_geom_line}
\alias{factory_geom_line}
\title{Create an alternative color aesthetic}
\usage{
factory_geom_line(aes_name)
}
\arguments{
\item{aes_name}{string skalar, the name of the new color aesthetic.}
}
\description{
Create an alternative color aesthetic
}
\examples{
factory_geom_line('var2')

}
