% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_factory_point.R
\name{factory_geom_point}
\alias{factory_geom_point}
\title{Create an alternative color aesthetic for geom_point}
\usage{
factory_geom_point(aes_name)
}
\arguments{
\item{aes_name}{string skalar, the name of the new geom}
}
\description{
Create an alternative color aesthetic for geom_point
}
\details{
The newly created aesthetics are then \code{test_c} (color) and \code{test_f} (fill).
}
\examples{
factory_geom_point('var3')

}
