% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_factory_rect.R
\name{factory_geom_rect}
\alias{factory_geom_rect}
\title{Create an alternative color aesthetic for geom_rect}
\usage{
factory_geom_rect(aes_name)
}
\arguments{
\item{aes_name}{string skalar, the name of the new geom}
}
\description{
Create an alternative color aesthetic for geom_rect
}
\details{
The newly created aesthetics are then \code{test_c} (color) and \code{test_f} (fill).
}
\examples{
factory_geom_rect('var3')

}
